+++ packages/jsenv-plugin-react/tests/react_refresh/react_refresh.test_manual.mjs
@@ -11,7 +11,7 @@ await startDevServer({
   certificate,
   privateKey,
   rootDirectoryUrl: new URL("./client/", import.meta.url),
-  plugins: [jsenvPluginReact()],
+  plugins: [jsenvPluginReact({ refresh: true })],
   explorer: {
     groups: {
       main: {