+++ src/kitchen/kitchen.js
@@ -38,7 +38,7 @@ export const createKitchen = ({
   sourcemaps = scenarios.dev ? "inline" : "none", // "programmatic" and "file" also allowed
   sourcemapsSourcesProtocol,
   sourcemapsSourcesContent,
-  sourcemapsRelativeSources,
+  sourcemapsSourcesRelative,
   writeGeneratedFiles,
   outDirectoryUrl,
 }) => {
@@ -263,7 +263,7 @@ ${ANSI.color(normalizedReturnValue, ANSI.YELLOW)}
     sourcemaps,
     sourcemapsSourcesProtocol,
     sourcemapsSourcesContent,
-    sourcemapsRelativeSources,
+    sourcemapsSourcesRelative,
     clientRuntimeCompat,
     injectSourcemapPlaceholder: ({ urlInfo, specifier }) => {
       const [sourcemapReference, sourcemapUrlInfo] = resolveReference(
@@ -652,6 +652,10 @@ ${ANSI.color(normalizedReturnValue, ANSI.YELLOW)}
       if (generatedUrl && generatedUrl.startsWith("file:")) {
         if (urlInfo.type === "directory") {
           // no need to write the directory
+        } else if (urlInfo.content === null) {
+          // Some error might lead to urlInfo.content to be null
+          // (error hapenning before urlInfo.content can be set, or 404 for instance)
+          // in that case we can't write anything
         } else {
           writeFileSync(new URL(generatedUrl), urlInfo.content)
           const { sourcemapGeneratedUrl, sourcemap } = urlInfo
@@ -847,8 +851,14 @@ const adjustUrlSite = (urlInfo, { urlGraph, url, line, column }) => {
         isOriginal: true,
         url: inlineUrlSite.url,
         content: inlineUrlSite.content,
-        line: inlineUrlSite.line + urlSite.line,
-        column: inlineUrlSite.column + urlSite.column,
+        line:
+          inlineUrlSite.line === undefined
+            ? urlSite.line
+            : inlineUrlSite.line + urlSite.line,
+        column:
+          inlineUrlSite.column === undefined
+            ? urlSite.column
+            : inlineUrlSite.column + urlSite.column,
       },
       parentUrlInfo,
     )